C1: engine scaffold — add engine alias, move rng to src/engine/core
Add 'engine' to the vite alias regex (covers vite, vitest, and tsc via
baseUrl). Move rng.ts + rng.test.ts verbatim from components/ephemeris
to src/engine/core, update the three importers, and add the
src/engine/index.ts public-API barrel starting with the rng exports.
