+++ src/components/ephemeris/sectorContent.test.ts
@@ -1,7 +1,7 @@
 // @vitest-environment happy-dom
 import * as THREE from 'three';
 import { describe, expect, it } from 'vitest';
-import { hashCoords, mulberry32 } from 'components/ephemeris/rng';
+import { hashCoords, mulberry32 } from 'engine/core/rng';
 import { buildHomeSystem, buildSectorContent } from 'components/ephemeris/sectorContent';
 
 const SECTOR = 700;
