+++ src/engine/core/rng.ts
