+++ src/engine/index.ts
@@ -1,0 +1,6 @@
+/**
+ * Shared space-engine core used by both the DEEP FIELD landing background
+ * and the EPHEMERIS game. Public API barrel — grows as engine modules land.
+ */
+
+export { hashCoords, makeName, mulberry32, pickFrom, SYLLABLES } from 'engine/core/rng';
